requirements.txt
Django==5.2.1
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.2
celery>=5.4.0
redis>=4.5.0
django-celery-beat>=2.6.0
gunicorn
python-dotenv
stripe
